signup and component: creating component DatePicker and updating it in singup
